DataType externalized to a single .hpp file
